requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
python-whois>=0.9.3
scikit-learn>=1.4.0
